Update README.md
We discussed about the Linux that how it is useful in programing language.It is accessible,and has lots of features that a user can use based on its wish.It is an open source operating system.In one of the class, we downloaded it from the website and then we started working on it.First we set the password for our use and we got to know the use of powershell,commands shell and so on.We can give commands on commands shell to work on our files.We can open vlc mediaplayer through the commands on this shell without opening it directly.Any files can be opened.Thta is why it is used for the database management syste.We can transfer one texts of file to other files.We can also list out the number of files created in the history.Working on the files are like,
1)First we create a directory(mkdir) to work.
2)giving commands like ls(to list out the files that we have)
3)to move files we use the command, cd.
4)pico command for the gaining the permission for writing.
5)rm command is used to remove or delete the folder and so on.
These are just examples that what we have learnt in this linux class.We learnt the following things like how to copy files from one directory to another directory,how to collect all files in one file,how to restart the pc through the data stored in the pendrive. and so on.The working on programming means that how communicating with machine.The two things I was fascinated about was that a route opened in between the start of the laptop before the starting of the display of the pc with icons.It shows that it was the detail of the pc that holded and we entered into the middle of the start of the operating system and closing of the operating system.And second thing was opening of any file or vlc media player with the use of coding in linux without  going directory to the file.We had done the same activity with pictures, word texts and audios.It new and fun to learn it.
One example for opening a file through commands is that,
1)Go to file of chrome
2) Copy the location path
3)And paste it in command propt.
4)Then,write”DIR” and press ENTER 
5)And then give command “chrome.exe”
The chrome will be opened.So, this is how it works through the commands.This exercise fascinated me a lot.